[hir2mpl]Internal commit msg:     [codecheck ]     [codecheck]     [hir2mpl]support stmt memberexpr [mapleall]Internal commit msg:     [codecheck ra emit isel cgfunc peep]     [codecheck cg]     [ICSL] Clean Code 37     [fix IsOperandImmValid memLeak]     [mpl2mpl] skip struct-embedded array when replace const     [driver]support pgo     [mapleall] cleancode2  40     fix const compute bug     [mplcg] fix maple string print bug     cg_bugfix_fix_ret_bb_live_out_reg     cg_bugfix_fix_global_struct_f128_align     [cg]Bugfix for DTS2023042803990     skip float loop test in lfo     [mplcg] Enhance peephole LsrAndToUbfxPattern     [mplcg] Address offset fold with pre&post index memopnd     [virtual sign]     [codecheck ]     [simplify] deal with struct-embedded array when replace const     [codecheck emit-cpp chainout-cpp]     [lexer, parser] int128: Support int128 constants in lexer/parser     [ICSL] Clean Code 36     [ICSL] Clean Code 35     [ICSL] Clean Code 34     The zero-witdth bit field is not affected by attr packed.     [me] reduce maydef/use for functions without side effect     [codeclean cg]     [mplcg] Bugfix forDTS2023042508040 & DTS2023042508043     [ICSL] Clean Code 33     [mplcg] Bugfix for DTS2023042405242     [mapleall] bugfix find element in map     [abi] hva/hfa check struct has union     eval (dread agg) case should return     [codecheck sim cg]     [cg]Bugfix for DTS2023042508045     [ICSL] Clean Code 32     validbit bugfix     [mapleall] bugfix compiler err in phase vrp     [driver]fixbug in ld_libary_path     [driver]ld option -l     [codecheck]     [cg]Bugfix for DTS2023042402188     [mapleall] cleancode2  38     [mapleall] bugfix vrp with delete assign to global var     [mapleall] bugfix vrp when deal with switch stmt     [codecheck extern includeh cpctor]     cg_bugfix_fix_unreachable_bb_caused_by_dead_loop     cg_feature_not_emit_debuginfo_in_debug_mode     [driver]fpic     fix dom's validity maintenance     [mplcg] Bugfix for DTS2023041500602     sra clean code     cg_bugfix_adjust_empty_bb_opt_for_dbg_insn     [hpf] remove unnecessary libc symbols     [mplcg]fix pgo use(3)

+++ src/mapleall/maple_be/include/cg/operand.h
@@ -470,7 +470,7 @@ class ImmOperand : public OperandVisitable<ImmOperand> {
     return relocs;
   }
 
-  bool IsInBitSize(uint8 size, uint8 nLowerZeroBits) const {
+  bool IsInBitSize(uint32 size, uint32 nLowerZeroBits) const {
     return IsBitSizeImmediate(static_cast<uint64>(value), size, nLowerZeroBits);
   }
